matrix_pickle

Trait Encode

pub trait Encode {
    // Required method
    fn encode(&self, writer: &mut impl Write) -> Result<usize, EncodeError>;

    // Provided method
    fn encode_to_vec(&self) -> Result<Vec<u8>, EncodeError> { ... }
}

A trait for encoding values into the matrix-pickle binary format.

Required Methods

fn encode(&self, writer: &mut impl Write) -> Result<usize, EncodeError>

Try to encode and write a value to the given writer, returning how many bytes were written.

Provided Methods

fn encode_to_vec(&self) -> Result<Vec<u8>, EncodeError>

Try to encode a value into a new Vec.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.